feat: add pydantic launcher config model
