+++ launcher_core/config/load_launcher_config.py
@@ -19,27 +19,8 @@ except ModuleNotFoundError:  # pragma: no cover - Py311 ships tomllib
     import tomli as tomllib
 
 from tomli_w import dumps
-from pydantic import BaseModel, Field
 
-
-class LauncherConfig(BaseModel):
-    """Minimal launcher configuration used by the test-suite.
-
-    Only the options that are referenced in the tests are implemented
-    here.  Defaults mirror the expectations expressed in the tests.
-    """
-
-    launcher_name: str = "AsyncMCLauncher"
-    launcher_version: str = "1.0.0"
-    username: Optional[str] = None
-    version: Optional[str] = None
-    concurrent_downloads: int = Field(4, ge=1, le=16)
-    download_timeout: int = Field(300, ge=1)
-    verify_downloads: bool = True
-    auto_refresh_token: bool = True
-    log_level: str = "INFO"
-    resolution_width: Optional[int] = None
-    resolution_height: Optional[int] = None
+from launcher_core.models import LauncherConfig
 
 
 class ConfigManager:
